qemu: add hard reboot in QEMU driver
Add the new flag VIR_DOMAIN_REBOOT_HARD/VIR_DOMAIN_SHUTDOWN_HARD to
carry out a hard reboot, which kills the QEMU process and creates a new
one with the same definition.

Hard reboot will be the highest priority to check. If succeed, other
reboot policy (i.e. agent and acpi) would be skipped. Otherwise, use the
traditional way.

To make the shutdown graceful, the workflow of hard reboot is similar
to the existing fakeReboot procedure.

 - In qemuDomainObjPrivatePtr we have a bool hardReboot flag.
 - The virDomainReboot method sets this flag and then triggers a normal
   "system_powerdown" if we specify the reboot mode "--mode hard"
 - The QEMU process is started with '-no-shutdown' so that the guest
   CPUs pause when it powers off the guest.
 - When we receive the "SHUTDOWN" event from QEMU monitor, if hardReboot
   is not set, then check fakeReboot or the last choice to invoke
   qemuProcessKill command to shutdown normally.
 - If hardReboot was set, we spawn a background thread, which issues
   qemuProcessStop to kill the QEMU process and cleanup the working
   environment. Then issue qemuProcessStart to create a new QEMU process
   with the same domain definition. At last, issues 'cont' to start the
   CPUs again.

The state transfer during the hardReboot is
RUNNING->SHUTDOWN->SHUTOFF->PAUSED->RUNNING, This patch doesn't hide
it as the states is not transient.

+++ src/conf/domain_conf.c
@@ -13498,7 +13498,7 @@ static int
 virDomainTDXDefParseXML(virDomainTDXDef *def,
                         xmlXPathContextPtr ctxt)
 {
-    unsigned long policy;
+    unsigned long long policy;
 
     if (virXPathULongLongBase("string(./policy)", ctxt, 16, &policy) < 0) {
         virReportError(VIR_ERR_XML_ERROR, "%s",
